+++ packages/flutter_bloc/lib/src/bloc_listener.dart
@@ -23,15 +23,15 @@ typedef BlocWidgetListener<S> = void Function(BuildContext context, S state);
 typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 
 /// {@template bloc_listener}
-/// Takes a [BlocWidgetListener] and an optional [cubit] and invokes
-/// the [listener] in response to `state` changes in the [cubit].
+/// Takes a [BlocWidgetListener] and an optional [value] and invokes
+/// the [listener] in response to `state` changes in the [value].
 /// It should be used for functionality that needs to occur only in response to
 /// a `state` change such as navigation, showing a `SnackBar`, showing
 /// a `Dialog`, etc...
 /// The [listener] is guaranteed to only be called once for each `state` change
 /// unlike the `builder` in `BlocBuilder`.
 ///
-/// If the [cubit] parameter is omitted, [BlocListener] will automatically
+/// If the [value] parameter is omitted, [BlocListener] will automatically
 /// perform a lookup using [BlocProvider] and the current `BuildContext`.
 ///
 /// ```dart
@@ -42,12 +42,12 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 ///   child: Container(),
 /// )
 /// ```
-/// Only specify the [cubit] if you wish to provide a [cubit] that is otherwise
+/// Only specify the [value] if you wish to provide a bloc/cubit that is otherwise
 /// not accessible via [BlocProvider] and the current `BuildContext`.
 ///
 /// ```dart
 /// BlocListener<BlocA, BlocAState>(
-///   bloc: blocA,
+///   value: blocA,
 ///   listener: (context, state) {
 ///     // do stuff here based on BlocA's state
 ///   },
@@ -59,11 +59,11 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 /// {@template bloc_listener_listen_when}
 /// An optional [listenWhen] can be implemented for more granular control
 /// over when [listener] is called.
-/// [listenWhen] will be invoked on each [cubit] `state` change.
+/// [listenWhen] will be invoked on each `state` change.
 /// [listenWhen] takes the previous `state` and current `state` and must
 /// return a [bool] which determines whether or not the [listener] function
 /// will be invoked.
-/// The previous `state` will be initialized to the `state` of the [cubit]
+/// The previous `state` will be initialized to the `state` of the bloc/cubit
 /// when the [BlocListener] is initialized.
 /// [listenWhen] is optional and if omitted, it will default to `true`.
 ///
@@ -80,13 +80,13 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 /// )
 /// ```
 /// {@endtemplate}
-class BlocListener<C extends Cubit<S>, S> extends BlocListenerBase<C, S>
+class BlocListener<T extends Cubit<S>, S> extends BlocListenerBase<T, S>
     with BlocListenerSingleChildWidget {
   /// {@macro bloc_listener}
   const BlocListener({
     Key key,
     @required BlocWidgetListener<S> listener,
-    C cubit,
+    T value,
     BlocListenerCondition<S> listenWhen,
     Widget child,
   })  : assert(listener != null),
@@ -94,25 +94,25 @@ class BlocListener<C extends Cubit<S>, S> extends BlocListenerBase<C, S>
           key: key,
           child: child,
           listener: listener,
-          cubit: cubit,
+          value: value,
           listenWhen: listenWhen,
         );
 }
 
 /// {@template bloc_listener_base}
-/// Base class for widgets that listen to state changes in a specified [cubit].
+/// Base class for widgets that listen to state changes in a specified bloc/cubit.
 ///
 /// A [BlocListenerBase] is stateful and maintains the state subscription.
 /// The type of the state and what happens with each state change
 /// is defined by sub-classes.
 /// {@endtemplate}
-abstract class BlocListenerBase<C extends Cubit<S>, S>
+abstract class BlocListenerBase<T extends Cubit<S>, S>
     extends SingleChildStatefulWidget {
   /// {@macro bloc_listener_base}
   const BlocListenerBase({
     Key key,
     this.listener,
-    this.cubit,
+    this.value,
     this.child,
     this.listenWhen,
   }) : super(key: key, child: child);
@@ -121,9 +121,9 @@ abstract class BlocListenerBase<C extends Cubit<S>, S>
   /// [BlocListenerBase].
   final Widget child;
 
-  /// The [cubit] whose `state` will be listened to.
-  /// Whenever the [cubit]'s `state` changes, [listener] will be invoked.
-  final C cubit;
+  /// The bloc/cubit which will be listened to.
+  /// Whenever the `state` changes, [listener] will be invoked.
+  final T value;
 
   /// The [BlocWidgetListener] which will be called on every `state` change.
   /// This [listener] should be used for any code which needs to execute
@@ -134,34 +134,34 @@ abstract class BlocListenerBase<C extends Cubit<S>, S>
   final BlocListenerCondition<S> listenWhen;
 
   @override
-  SingleChildState<BlocListenerBase<C, S>> createState() =>
-      _BlocListenerBaseState<C, S>();
+  SingleChildState<BlocListenerBase<T, S>> createState() =>
+      _BlocListenerBaseState<T, S>();
 }
 
-class _BlocListenerBaseState<C extends Cubit<S>, S>
-    extends SingleChildState<BlocListenerBase<C, S>> {
+class _BlocListenerBaseState<T extends Cubit<S>, S>
+    extends SingleChildState<BlocListenerBase<T, S>> {
   StreamSubscription<S> _subscription;
   S _previousState;
-  C _cubit;
+  T _bloc;
 
   @override
   void initState() {
     super.initState();
-    _cubit = widget.cubit ?? context.read<C>();
-    _previousState = _cubit.state;
+    _bloc = widget.value ?? context.read<T>();
+    _previousState = _bloc.state;
     _subscribe();
   }
 
   @override
-  void didUpdateWidget(BlocListenerBase<C, S> oldWidget) {
+  void didUpdateWidget(BlocListenerBase<T, S> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    final oldCubit = oldWidget.cubit ?? context.read<C>();
-    final currentCubit = widget.cubit ?? oldCubit;
-    if (oldCubit != currentCubit) {
+    final oldBloc = oldWidget.value ?? context.read<T>();
+    final currentBloc = widget.value ?? oldBloc;
+    if (oldBloc != currentBloc) {
       if (_subscription != null) {
         _unsubscribe();
-        _cubit = currentCubit;
-        _previousState = _cubit.state;
+        _bloc = currentBloc;
+        _previousState = _bloc.state;
       }
       _subscribe();
     }
@@ -177,8 +177,8 @@ class _BlocListenerBaseState<C extends Cubit<S>, S>
   }
 
   void _subscribe() {
-    if (_cubit != null) {
-      _subscription = _cubit.listen((state) {
+    if (_bloc != null) {
+      _subscription = _bloc.listen((state) {
         if (widget.listenWhen?.call(_previousState, state) ?? true) {
           widget.listener(context, state);
         }