+++ packages/flutter_bloc/test/bloc_listener_test.dart
@@ -38,7 +38,7 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         body: BlocListener<CounterCubit, int>(
-          cubit: _counterCubit,
+          value: _counterCubit,
           listener: (context, state) {
             widget.onListenerCalled?.call(context, state);
           },
@@ -70,12 +70,12 @@ class _MyAppState extends State<MyApp> {
 
 void main() {
   group('BlocListener', () {
-    testWidgets('throws if initialized with null cubit, listener, and child',
+    testWidgets('throws if initialized with null value, listener, and child',
         (tester) async {
       try {
         await tester.pumpWidget(
           BlocListener<Cubit, dynamic>(
-            cubit: null,
+            value: null,
             listener: null,
             child: null,
           ),
@@ -91,7 +91,7 @@ void main() {
       try {
         await tester.pumpWidget(
           BlocListener<CounterCubit, int>(
-            cubit: CounterCubit(),
+            value: CounterCubit(),
             listener: null,
             child: null,
           ),
@@ -106,7 +106,7 @@ void main() {
       const targetKey = Key('cubit_listener_container');
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: CounterCubit(),
+          value: CounterCubit(),
           listener: (_, __) {},
           child: const SizedBox(key: targetKey),
         ),
@@ -120,7 +120,7 @@ void main() {
       const expectedStates = [1];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listener: (_, state) {
             states.add(state);
           },
@@ -138,7 +138,7 @@ void main() {
       const expectedStates = [1, 2];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listener: (_, state) {
             states.add(state);
           },
@@ -234,7 +234,7 @@ void main() {
       const expectedStates = [1];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (previous, state) {
             listenWhenCallCount++;
             latestPreviousState = previous;
@@ -263,7 +263,7 @@ void main() {
       const expectedStates = [2];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (previous, state) {
             listenWhenCallCount++;
             if ((previous + state) % 3 == 0) {
@@ -293,7 +293,7 @@ void main() {
       final counterCubit = CounterCubit();
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (previous, current) {
             if (current % 3 == 0) {
               listenWhenPreviousState.add(previous);
@@ -355,7 +355,7 @@ void main() {
       const expectedStates = [1, 2];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (previous, state) {
             listenWhenCallCount++;
             latestPreviousState = previous;
@@ -385,7 +385,7 @@ void main() {
       const expectedStates = <int>[];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (_, __) => false,
           listener: (_, state) => states.add(state),
           child: const SizedBox(),
@@ -405,7 +405,7 @@ void main() {
       const expectedStates = [1];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (_, __) => true,
           listener: (_, state) => states.add(state),
           child: const SizedBox(),
@@ -425,7 +425,7 @@ void main() {
       const expectedStates = <int>[];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (_, __) => false,
           listener: (_, state) => states.add(state),
           child: const SizedBox(),
@@ -451,7 +451,7 @@ void main() {
       const expectedStates = [1, 2, 3, 4];
       await tester.pumpWidget(
         BlocListener<CounterCubit, int>(
-          cubit: counterCubit,
+          value: counterCubit,
           listenWhen: (_, __) => true,
           listener: (_, state) => states.add(state),
           child: const SizedBox(),